# gateway/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MatchRequest": {
        "description": "Request model for map matching a sequence of GPS breadcrumbs.",
        "examples": [
          {
            "breadcrumbs": [
              {
                "accuracy_meters": 5.0,
                "latitude": 9.9281,
                "longitude": -84.0907,
                "timestamp": 1713000000
              },
              {
                "accuracy_meters": 5.0,
                "latitude": 9.93,
                "longitude": -84.088,
                "timestamp": 1713000030
              },
              {
                "accuracy_meters": 5.0,
                "latitude": 9.9333,
                "longitude": -84.0833,
                "timestamp": 1713000060
              }
            ],
            "geometries": "geojson",
            "profile": "driving",
            "tidy": true
          }
        ],
        "properties": {
          "annotations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Comma-separated metadata annotations per segment",
            "title": "Annotations"
          },
          "approaches": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "enum": [
                        "unrestricted",
                        "curb"
                      ],
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate approach side. Use null for default.",
            "title": "Approaches"
          },
          "bearings": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate bearing constraints as 'angle,deviation' strings (e.g. '90,30'). Use null for unconstrained coordinates.",
            "title": "Bearings"
          },
          "breadcrumbs": {
            "items": {
              "$ref": "#/components/schemas/GPSBreadcrumb"
            },
            "maxItems": 5000,
            "minItems": 2,
            "title": "Breadcrumbs",
            "type": "array"
          },
          "exclude": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Road classes to exclude globally (e.g. ['motorway', 'toll']).",
            "title": "Exclude"
          },
          "gaps": {
            "anyOf": [
              {
                "enum": [
                  "split",
                  "ignore"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "How to handle large timestamp gaps in the trace",
            "title": "Gaps"
          },
          "geometries": {
            "default": "geojson",
            "description": "Geometry encoding format",
            "enum": [
              "polyline",
              "polyline6",
              "geojson"
            ],
            "title": "Geometries",
            "type": "string"
          },
          "hints": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate hint strings from a previous OSRM response to speed up snapping.",
            "title": "Hints"
          },
          "match_waypoints": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Indices into breadcrumbs to treat as explicit waypoints",
            "title": "Match Waypoints"
          },
          "overview": {
            "default": "full",
            "description": "Level of overview geometry returned",
            "enum": [
              "simplified",
              "full",
              "false"
            ],
            "title": "Overview",
            "type": "string"
          },
          "profile": {
            "default": "driving",
            "description": "Routing profile",
            "enum": [
              "driving",
              "cycling",
              "walking"
            ],
            "title": "Profile",
            "type": "string"
          },
          "radiuses": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate snapping radius in meters. Use null for unlimited.",
            "title": "Radiuses"
          },
          "skip_waypoints": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Suppress the waypoints array in the response to reduce payload size.",
            "title": "Skip Waypoints"
          },
          "snapping": {
            "anyOf": [
              {
                "enum": [
                  "default",
                  "any"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Edge selection behaviour: 'default' (one-way roads respected) or 'any'.",
            "title": "Snapping"
          },
          "steps": {
            "default": false,
            "description": "Return step-by-step instructions for the matched route",
            "title": "Steps",
            "type": "boolean"
          },
          "tidy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Remove repeated or out-of-order coordinates before matching",
            "title": "Tidy"
          }
        },
        "required": [
          "breadcrumbs"
        ],
        "title": "MatchRequest",
        "type": "object"
      },
      "MatrixRequest": {
        "description": "Request model for generating a distance matrix.",
        "examples": [
          {
            "annotations": "duration,distance",
            "coordinates": [
              {
                "latitude": 9.9281,
                "longitude": -84.0907
              },
              {
                "latitude": 9.9333,
                "longitude": -84.0833
              },
              {
                "latitude": 9.94,
                "longitude": -84.1
              }
            ],
            "profile": "driving"
          }
        ],
        "properties": {
          "annotations": {
            "default": "duration,distance",
            "description": "Which cost annotations to return",
            "enum": [
              "duration",
              "distance",
              "duration,distance"
            ],
            "title": "Annotations",
            "type": "string"
          },
          "approaches": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "enum": [
                        "unrestricted",
                        "curb"
                      ],
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate approach side. Use null for default.",
            "title": "Approaches"
          },
          "bearings": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate bearing constraints as 'angle,deviation' strings (e.g. '90,30'). Use null for unconstrained coordinates.",
            "title": "Bearings"
          },
          "coordinates": {
            "items": {
              "$ref": "#/components/schemas/Coordinate"
            },
            "maxItems": 5000,
            "minItems": 2,
            "title": "Coordinates",
            "type": "array"
          },
          "destinations": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destinations"
          },
          "exclude": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Road classes to exclude globally (e.g. ['motorway', 'toll']).",
            "title": "Exclude"
          },
          "fallback_coordinate": {
            "anyOf": [
              {
                "enum": [
                  "input",
                  "snapped"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which coordinate to use when a pair falls back to speed estimate",
            "title": "Fallback Coordinate"
          },
          "fallback_speed": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Speed (km/h) used to estimate travel time for unreachable pairs",
            "title": "Fallback Speed"
          },
          "hints": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate hint strings from a previous OSRM response to speed up snapping.",
            "title": "Hints"
          },
          "profile": {
            "default": "driving",
            "description": "Routing profile",
            "enum": [
              "driving",
              "cycling",
              "walking"
            ],
            "title": "Profile",
            "type": "string"
          },
          "radiuses": {
            "anyOf": [
              {
                "items": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-coordinate snapping radius in meters. Use null for unlimited.",
            "title": "Radiuses"
          },
          "scale_factor": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Multiply all durations by this factor (useful for congestion modelling)",
            "title": "Scale Factor"
          },
          "skip_waypoints": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Suppress the waypoints array in the response to reduce payload size.",
            "title": "Skip Waypoints"
          },
          "snapping": {
            "anyOf": [
              {
                "enum": [
                  "default",
                  "any"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Edge selection behaviour: 'default' (one-way roads respected) or 'any'.",
            "title": "Snapping"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "MatrixRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VrpAllocationResponse": {
        "description": "Response model for the Location-Allocation (Clustering) phase.",
        "properties": {
          "allocations": {
            "additionalProperties": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Allocations",
            "type": "object"
          },
          "code": {
            "default": "Ok",
            "title": "Code",
            "type": "string"
          },
          "unreachable_stops": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Unreachable Stops",
            "type": "array"
          }
        },
        "required": [
          "allocations",
          "unreachable_stops"
        ],
        "title": "VrpAllocationResponse",
        "type": "object"
      },
...
      "VrpResponse": {
        "description": "Response model for the VRP solution.",
        "properties": {
          "code": {
            "default": "Ok",
            "title": "Code",
            "type": "string"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/VehicleRoute"
            },
            "title": "Routes",
            "type": "array"
          },
          "total_distance": {
            "title": "Total Distance",
            "type": "number"
          },
          "total_duration": {
            "title": "Total Duration",
            "type": "number"
          }
        },
        "required": [
          "routes",
          "total_distance",
          "total_duration"
        ],
        "title": "VrpResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OSRM API Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/match": {
      "post": {
        "description": "Map match noisy GPS traces to the road network.\nHandles multiple segments automatically (Trace Splitting).",
        "operationId": "match_trace_match_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Match Trace Match Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Map Match GPS Traces",
        "tags": [
          "Map Matching"
        ]
      }
    },
    "/matrix": {
      "post": {
        "description": "Fetch raw distance/duration matrix.",
        "operationId": "get_matrix_matrix_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatrixRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Matrix Matrix Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Distance/Duration Matrix",
        "tags": [
          "Matrix"
        ]
      }
    },
    "/matrix-graph": {
      "post": {
        "description": "Generate a directed graph from a distance/duration matrix.",
        "operationId": "get_matrix_graph_matrix_graph_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatrixRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Matrix Graph Matrix Graph Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Matrix as Graph",
        "tags": [
          "Matrix"
        ]
      }
    },
...
    "/nearest": {
      "post": {
        "description": "Find the nearest road network node(s) to a given coordinate.\nUseful for snapping raw GPS coordinates to the routable road graph.",
        "operationId": "get_nearest_nearest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NearestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Nearest Nearest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Snap Coordinate to Road Network",
        "tags": [
          "Routing"
        ]
      }
    },
...
    "/route": {
      "post": {
        "description": "Calculate highly accurate driving route.",
        "operationId": "get_route_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Route Route Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Route",
        "tags": [
          "Routing"
        ]
      }
    },
...
    "/trip": {
      "post": {
        "description": "Solve TSP for an optimized sequence of coordinates.\nIdeal for comparing actual vs optimized routes.",
        "operationId": "get_trip_trip_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Trip Trip Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Solve Traveling Salesperson Problem",
        "tags": [
          "Optimization"
        ]
      }
    },
    "/vrp": {
      "post": {
        "description": "Solve multi-vehicle Vehicle Routing Problem using Location-Allocation.\nAssigns stops to the logistically closest warehouse before routing.\n\nArgs:\n    request: Incoming request, used by the rate limiter.\n    payload: Depots, stops, and clustering options for the solve.\n\nReturns:\n    VrpResponse: One optimized route per vehicle.\n\nRaises:\n    HTTPException: 503 when solve capacity is exhausted, 500 on failure.",
        "operationId": "solve_vrp_vrp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VrpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VrpResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Solve Vehicle Routing Problem",
        "tags": [
          "Optimization"
        ]
      }
    },
    "/vrp/allocate": {
      "post": {
        "description": "Cluster products to warehouses using Location-Allocation.\nThis is the first phase of the VRP solving process.\nUseful for seeing product-to-depot assignments before routing.\n\nArgs:\n    request: Incoming request, used by the rate limiter.\n    payload: Depots, stops, and clustering options for the allocation.\n\nReturns:\n    VrpAllocationResponse: Depot-to-stop assignments and unreachable stops.\n\nRaises:\n    HTTPException: 503 when solve capacity is exhausted, 500 on failure.",
        "operationId": "allocate_vrp_vrp_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VrpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VrpAllocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cluster Products to Warehouses",
        "tags": [
          "Optimization"
        ]
      }
    }
  }
}